LayoutTests/fast/events/mouse-click-events.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This tests what mouse events we send.");

var div = document.createElement("div");
div.style.width = "100px";
div.style.height = "100px";
div.style.backgroundColor = "blue";

var eventLog = "";

function appendEventLog() {
    if (window.eventSender) {
        eventLog += event.type + "(" + event.button + ") ";
    } else {
        debug(event.type + "(" + event.button + ")");
    }
}

function clearEventLog() {
    eventLog = "";
}

function dismissContextMenu() {
    if (window.eventSender) {
        // esc key to kill the context menu.
        eventSender.keyDown(String.fromCharCode(0x001B), null);
    }
}

async function sendEvents(button) {
    if (!window.eventSender) {
        debug("This test requires DumpRenderTree.  Click on the blue rect with different mouse buttons to log.")
        return;
    }
    await eventSender.asyncMouseDown(button);
    if (button == 2)
        dismissContextMenu();
    await eventSender.asyncMouseUp(button);
    await eventSender.asyncMouseDown(button);
    if (button == 2)
        dismissContextMenu();
    await eventSender.asyncMouseUp(button);
    // could test dragging here too
}

async function testEvents(description, button, expectedString) {
    debug(description);
    await sendEvents(button);
    shouldBeEqualToString("eventLog", expectedString);
    clearEventLog();
}

onload = async () => {
    testRunner?.waitUntilDone();

    div.addEventListener("auxclick", appendEventLog, false);
    div.addEventListener("click", appendEventLog, false);
    div.addEventListener("dblclick", appendEventLog, false);
    div.addEventListener("mousedown", appendEventLog, false);
    div.addEventListener("mouseup", appendEventLog, false);
    document.body.insertBefore(div, document.body.firstChild);

    if (window.eventSender) {
        await eventSender.asyncMouseMoveTo(10, 10);
        await testEvents("Left Mouse Button", 0, "mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) dblclick(0) ");
        await testEvents("Middle Mouse Button", 1, "mousedown(1) mouseup(1) auxclick(1) mousedown(1) mouseup(1) auxclick(1) ");
        await testEvents("Right Mouse Button", 2, "mousedown(2) mouseup(2) auxclick(2) mousedown(2) mouseup(2) auxclick(2) ");
        await testEvents("4th Mouse Button", 3, "mousedown(1) mouseup(1) auxclick(1) mousedown(1) mouseup(1) auxclick(1) ");
    }
    testRunner?.notifyDone();
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
